refcator: move DTOs to separate files, add missing docs
