+++ MVPproject/MVPproject/Services/Network/DTO/DetailHitsDTO.swift
@@ -1,0 +1,10 @@
+// DetailHitsDTO.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// Блюдо-хит с деталями
+struct DetailHitsDTO: Codable {
+    /// рецепт
+    let recipe: DetailsDTO
+}
