+++ MVPproject/MVPproject/Services/Network/DTO/QuantityDTO.swift
@@ -1,0 +1,10 @@
+// QuantityDTO.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// Значения количества КБЖУ
+struct QuantityDTO: Codable {
+    /// Количество
+    let quantity: Double
+}
